requirements.txt
flask
anthropic
psycopg2-binary
gunicorn